requirements.txt
langchain>=0.2.3
langchain-core>=0.2.4
langchain-openai>=0.1.7
langchain-community>=0.2.2
pydantic>=2.7.0
python-dotenv>=1.0.1
typer[all]>=0.12.3
rich>=13.7.1
pytest>=8.2.0
ruff>=0.5.0
